backend/pyproject.toml
[project]
name = "ad-copy-backend"
version = "0.1.0"
description = "Ad Copy Dashboard Backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "supabase>=2.0.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.0.0",
]

[project.scripts]
dev = "uvicorn api:app --reload --host 0.0.0.0 --port 8000"
